BIG WIN [merge driver ul cum merge]

+++ EVR_VCU_K144/include/7-segment-display.h
@@ -86,7 +86,7 @@ typedef struct {
 ==================================================================================================*/
 
 void SevenSegmentInit(void);
-void SevenSegmentDisplayDecimalValue(uint8 SevenSegmentGroupIndex, uint8 DecimalValue, uint8 PrecisionFloatPoint);
+void SevenSegmentDisplayDecimalValue(uint8 SevenSegmentGroupIndex, uint16 DecimalValue, uint8 PrecisionFloatPoint);
 void SevenSegmentSetGlobalBrightness(uint8 BrightnessPercent);
 
 #ifdef __cplusplus
